split components by folders
