+++ src/components/pages/index/components/Canvas/index.ts
@@ -1,0 +1,3 @@
+import { Canvas } from './Canvas'
+
+export { Canvas }
