+++ src/components/pages/index/components/Options/components/RenderAction.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { useRecoilCallback } from 'recoil'
-import { backgroundBrightnessState, iterationsState } from '../state'
-import { CanvasContext } from '../context/CanvasContext'
-import { WASMContext } from '../context/WASMContext'
-import { WASMWorkerMessageType } from '../workers/wasm/types'
-import type { WASMWorkerRenderMessage } from '../workers/wasm/types'
-import { Button } from './Button'
+import { backgroundBrightnessState, iterationsState } from '@/state'
+import { CanvasContext } from '@/context/CanvasContext'
+import { WASMContext } from '@/context/WASMContext'
+import { WASMWorkerMessageType } from '@/workers/wasm/types'
+import type { WASMWorkerRenderMessage } from '@/workers/wasm/types'
+import { Button } from '@/components/ui/Button'
 
 export const RenderAction = () => {
   const canvasRef = React.useContext(CanvasContext)
