+++ src/components/pages/index/components/Options/index.ts
@@ -1,0 +1,3 @@
+import { Options } from './Options'
+
+export { Options }
